\encoding{utf8}
\name{ceram.n}
\docType{data}
\alias{ceram.n}
\title{Cerambycidae from Northern European Countries}
\description{
      The ceram.n data frame has 19 rows and 634 columns. Columns are presence/absence values of 634 species. The variable names are formed from the scientific names. The case names are standard country abbreviations, excepting RSS (Southern European Russia), RSC (Central European Russia) and RSN (Northern European Russia).
     }
\usage{data(ceram.n)}
\source{http://www.cerambycidae.net/}
\references{
\enumerate{
\item{Danilevsky, M. L. 2007. A check-list of Longicorn Beetles (Coleoptera, Cerambycoidea) of Europe.Available at http://www.cerambycidae.net/}
\item{Baselga, A. 2008. Determinants of species richness, endemism and turnover in European longhorn beetles. Ecography 31:263-271}
}}


