\encoding{utf8}
\name{ceram.s}
\alias{ceram.s}
\docType{data}
\title{Cerambycidae from Southern European Countries}
\description{
     The ceram.s data frame has 15 rows and 634 columns. Columns are presence/absence values of 634 species. The variable names are formed from the scientific names. The case names are standard country abbreviations, excepting SS (Serbia) and CBH (Croatia and Bosnia-Herzegovina).
     }
\usage{data(ceram.s)}
\source{http://www.cerambycidae.net/}
\references{
\enumerate{
\item{Danilevsky, M. L. 2007. A check-list of Longicorn Beetles (Coleoptera, Cerambycoidea) of Europe.Available at http://www.cerambycidae.net/}
\item{Baselga, A. 2008. Determinants of species richness, endemism and turnover in European longhorn beetles. Ecography 31:263-271}
}}





