\encoding{utf8}
\name{coords.n}
\alias{coords.n}
\docType{data}
\title{Spatial coordinates for Southern European Countries}
\description{
     The coords.n data frame has 18 rows and 4 columns. Columns are UTM and latlong coordinates. The row names are standard country abbreviations, excepting RSS (Southern European Russia), RSC (Central European Russia) and RSN (Northern European Russia).
     }
\usage{data(coords.n)}
\references{
\enumerate{
\item{Baselga, A. 2008. Determinants of species richness, endemism and turnover in European longhorn beetles. Ecography 31:263-271}
}}





