\encoding{utf8}
\name{coords.s}
\alias{coords.s}
\docType{data}
\title{Spatial coordinates for Southern European Countries}
\description{
     The coords.s data frame has 15 rows and 4 columns. Columns are UTM and latlong coordinates. TThe row names are standard country abbreviations, excepting SS (Serbia) and CBH (Croatia and Bosnia-Herzegovina).
     }
\usage{data(coords.s)}
\references{
\enumerate{
\item{Baselga, A. 2008. Determinants of species richness, endemism and turnover in European longhorn beetles. Ecography 31:263-271}
}}





